+++ docs/en/changes/changes.md
@@ -135,6 +135,7 @@
 * Add the spring-ai components and the GenAI layer.
 * Bump up netty to 4.2.10.Final.
 * Bump up log4j to 2.25.3 and jackson to 2.18.5.
+* Support TraceQL and Tempo API for Zipkin trace query.
 
 #### UI
 * Fix the missing icon in new native trace view.